Repository for Java
This repository was created as part of an evaluation of "Introduction to GIT"  on  Coursera. Later I decided to make use of the same repository to learn more things on Java and GIT. For that, I renamed this repository to Java_Stuffs. As part of this, I am reusing the README.md file for documentation purpose of the works I do on this repository